About to try and switch to CMake
-got the basics of json serialization and deserialization figured out for the engine, need to impement on the foundation i've created now

-created basic functionality to search for main working directory changes, so that the changes can be reflected accordingly in the editor in the case where assets are added outside of the editor environment. which is probably constantly, because i do that all the time LOL

-cleaned up the deprecated joint types form box2d 2.4.1, and updated to the new joint types in box2d 3.0.0 (RopeJoint always made me do a double take anyways so good riddance lmao)

-compiled GGPO from source, and statically linked it with the project. WE GOT ROLLBACK IN THE STARS FOR US BABY

I hope to god switching to CMake doesn't break everything lmao.

My main goal is to get the build system working on my laptop, and to also utilize CPack to create an installer for missing dll's that are required for running the editor and engine.

seems like we need to install some vc++ runtime dll's to allow our engine to run, along with python312.dll apparently lmao.

if all else fails ima just explicitly bundle the dll's with the editor, >w<