got files appearing
okay so the file system is working alright and you can select files and handle the selection through a function call, however its kinda jank when opening files and idk y

also its not very pretty and the examples on imgui arent promising.

so ima try to switch to nuklear after this commit

i think nuklear is a better choice since it seems light on deps, it abstracts to any rendering backend since it isnt backend dependent

AND

it seems like it will allow for more pretty graphics that we can use as a long-term solution, since the graphics will ultimately be drawn by the rendering code. So we can just change whatever design we design to render on top of the basic polygon output from nuklear from what i understand before reading the header atm

WISH ME LUCK FINGERS CROSSED OWO rawr