i forget but ima try including sfml src in the project
-bgfx is over complicated and the development process kinda blows,

i don't get why bgfx does shaders the way it does, it would make much more sense to bring the language up a level,

write the code in a new language, then compile to a taget shader lang

external weirdly created shaderc tool that was rekitted for bgfx does not cut it.

the weird bespoke build tool and all the weird libraries built around bgfx, it is just too much to be a viable tool for commericial projects in my opinion

if you're willing to deal with all this stupidity sure, but there are much easier ways too accomplishing the same end goal.

like the project feels like a ship of theseus  where he took libraries like bx and bimg and just built on it until he had his own "perfect" tool that he understood for doing stuff

it wasn't ever build from the groundup as a general tool for people and the lack of documentation shows that

it is just so insane that i can read so much src code and have the level of mathematics education that I do; and still be confused by this shit

im so sick of coders wanting to "feel cool" by writing incomprehensible garbage.

Like seriously, computers are only doing 86 instructions at the end of the fucking day, why are you including preprocessor garbage and weird development patterns for what IS a simple end goal.

every piece of source ive read to this point makin this engine, has been mostly garbage because the coding styles are so fucking out dated and overly verbose in terms of syntax and file structure management.

these old head cpp coders complain so much ab modern cpp and its over-abundance of features, yet all they do is use extra garbage in their code that makes it near unreadable and impossible to decipher what it means
unless i wanna take a two month crash course on how this retard who coded this dumbass codebase's mind labyrinth ticks.

AND ALL OF THIS TO JUSTIFY CODING IN A STYLE THEIR COMFORTABLE IN WITHOUT ADMITTING THEY DONT WANNA LEARN THE NEW FEATURES

ITS ALMOST LIKE THE NEW FEATURES MAKE CPP MUCH MORE SAFE AND ALSO TAKE AWAY STUPID SHIT

LIKE USING TEMPLATES INSTEAD OF DYNAMIC CASTS OOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOOO MUST BE REALLY SCARY FOR A 40 YR OLD PROGRAMMER TO HEAR

all of this retardation just to code in a way that is objectively inferior and more error-prone.

it's almost like the guy that invented the language, has made it better.
(HUGE SHOCK RIGHT)

#endrant