removed nlohmann json
-cereal seems sufficient for my use case so i removed nlohmann json

-refactored classes to accommodate for the removal of nlohmann json

i don't really need any dynamic json creation or compatibility with other libraries. I believe cereal should be more than sufficient since we'll be serializing to json in large well defined chunks because all the game engine structures are well defined in their size requirements and limitations and we wanna keep everything as rigid as possible so that static analysis of the code is easier.

each node has very well defined parameters and members and we will only be serializing whole peachnodes for project files and save data and deserializing for loading up a project in the editor or for save data if the dev wants to use json uwu.

we will be deserializing whole peachnode objects since only accessing certain parts of it isn't useful or recommended within the current architecture

less bloat = happier me