Improved overall architecture of engine and cleaned up some stuff
-fixed nuklear_impl_vulkan.h includes

-removed CreateMainSDLWindow() because it just felt redundant, so i rewokred CreateSDLWindow() to just do the same thing

-moved SDL_Init to PeachEngineManager because it just makes more sense to put into the InitializePeachEngine() method because it just makes more sense since PeachEditor and PeachEngine will call it for SDL setup

-removed hopefully the last reference to the windows specific "//" character for specifying directory inside the code used for intializing the font for nuklear

-moved CreateSDLWindow() to RenderingManager exclusively since PeachEditorRenderingManager doesn't need it as well uwu

-added vma to project as a header so no more static linking >w<

-renamed tests dir to docs because i wasnt even really using the tests directory for anything besides shaders so uh yeah probs should've put that into shaders to begin w lmfao

-updated readme with my intentions for making Peach-E

Still need to fix how the game instance is ran because right now its hella thread unsafe, also wanna move the InitializeOpenGL() method to RenderingManager, since it doesn't feel like it has to be duplicated.

I'm not against code duplication if it avoids weirdly complicated inheritance schemes or jank in general that i've seen in other codebases. Trying to plan for a "perfect" codebase that utilizes everything in the most efficient way possible off the bat is a recipe for hating coding and ruining a project before it even has the opportunity to get off the ground.

I'm more in the camp of, just get it working cleanly, and then refactor anything that is absurdly stupid. Much faster dev times with 95% less headaches and 10000% more fun.

remember kids, "premature optimization is the root of all evil"