Added better error handling and hopefully reduced security risks along with some other stuff
-Added exception handling to main.cpp for Peach_Editor and Peach_Engine to avoid any problems associated with calling destructors, eg. not calling close() on privileged file access that could lead to attack vectors for malicious actors

-Added a primitive segfault handler to ensure clean exit from program in the case of a dreaded segfault error >w<

-Blocked multi thread access to any LogManager instance, the error handling is kinda weird, and I'm not sure if i want an unauthorized use of a LogManager instance to crash the program. However, threading issues with internal loggers should only be caused from the cpp side of things, so any scripting done for a game shouldn't cause any errors since scripts execute in a sandboxed environment without access to the error handling internals of Peach-E/Peach_Editor. So throwing an error and stopping program execution with a console print should be sufficient for debugging Peach_Editor and Peach-E since these errors should only show up  during C++ development of core systems

-Removed fp_ThreadName parameter from LogAndPrint/Log and added it to the initialize() function since LogManager is designed for each instance to be accessed and used from the thread its been created on

-Started reworking the way game scenes are ran within Peach-E because the lazy thread creation within PeachEditorRenderingManager was kinda janky and not super thread safe. It's going to take some more work to get it to a point thats nicer however I think this is a better direction to go.

I'm going to figure out how I want scenes ran within the editor, because idk if utilizing the same renderer as the editor is a great way of doing things but w/e, I'd rather have a safe executing program than trying to squeeze every last drop of performance out while creating a game within the editor.

Also, any issues coming from a failed close() on a log file should hopefully be much harder with the tighter exception and error handling. I want Peach-E to just work, and not introduce various attack vectors for malicious actors into privileged parts of an end-user's OS. Of course with any software, by the nature of it just existing and running on hardware, it will introduce security risks, but that doesn't mean I want it too be easy.

My overall goal with Peach-E and Peach_Editor is to make a tool that works well, allows game makers to create beautiful things, and do so without having to worry about dumb things relating to the engine breaking or compromising the safety of gamer's computers. Just getting enough money to have a decent machine to play games is tough as is. Having that investment lead to having information extorted/extracted/erased is just fucked. Security flaws ruins the magic and beauty of computers which is just so fucking lame.

Everyone should be allowed to enjoy the insane achievement that is computers and the internet because that shit is just so cool, life is already hard enough.