PUI Updates
-Started really carving otu how ShapePrimitives should behave and figuring out responsibilities for how PUI is going to work

-Moved rendering backend specific stuff to their respective rendering backends, just makes more sense to have VulkanShaderAssets in the VulkanRenderer and vice versa with the OpenGLRenderer

-Moved the PeachUIManager -> VulkanRenderer, and gonna rework it into a generic PUI impl class, same for OpenGLRenderer

-Added all the basic UI building blocks I want/think should be sufficient to start building apps, im guessing any additions after this should just be refinement, reworking systems and adding features since from what i can tell most UI's are just a composition of these basic elements.

I'm taking the RISC approach to building menus, essentially im doing that with all of Peach-E

I feel like more basic general purpose parts, when allowed to be assembled together in complex ways gives much more freedom, and leads to more interesting results since u can pretty much build anything since the individual parts don't really assume much, or do much on their own

so u can chain a bunch of them together to get more fine grained control and detail when building UI's or a game via the scene tree elements.

Big parts are nice for prototyping i guess but u lose a lot of that control in the process. Since they paint in large strokes and make a lot of assumptions along the way

Simpler parts lead to a nicer design pattern in my opinion, even if they take a bit to figure out how to build the final product, since building a UI using PUI isnt like "okay so we create a scrollable text region w a slider in a single node"

it's more like "okay so i can make a panel, then embed a scrollable region with a label, and i can track the location of the scrollable region using a slider that i can align against the parent panel"

it's a bit more thinking but allows u to do whatever the fuck u want w the UI, like a triangle slider (whatever that looks like)

and i feel like it's not a huge ask, instead of just being given a certain format, u have to think about whether that's the end design goal of ur UI, and if it is, YOU get to choose to implement that and get a few more variables to tweak to make it a bit more personal even if following typical convention uwu

(Changes nOT Tested)